Database/Migrations fix. int switched to decimal.

+++ database/migrations/2024_07_09_073538_create__data_commission_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('Type');
             $table->string('Name');
             $table->text('Description');
-            $table->integer('Charges');
-            $table->integer('Amount');
+            $table->decimal('Charges');
+            $table->decimal('Amount');
             $table->timestamps();
         });
     }
